Change Projects Card and Fix XDuke Button
I replaced the card in the projects section and fixed its hover effect. Now there is an animation when hovering over it. However, the individual page for each project has not been created yet. I also fixed the button for XDuke, which now takes him to the home page of the website.